requirements.txt
langchain
sentence-transformers
pinecone-client
cohere